lunch: add a catch-all handler
The catch-all handler is not *really* a handler.

When none of the protocol, extension, or MIME rules match the file/URL
given, $XDG_CONFIG_HOME/lunch/unknown will be ran with the file/URL
given, the last handler ran, and the last handler preference searched
for passed as arguments.

This could be used to send notifications (ex. using notify-send(1))
about failed runs, or perhaps to pass off to another opener.